+++ R/RcppExports.R
@@ -49,3 +49,7 @@ check_cbrace_expr_parser <- function(str) {
     .Call(`_parsermd_check_cbrace_expr_parser`, str)
 }
 
+check_code_block_parser <- function(str) {
+    .Call(`_parsermd_check_code_block_parser`, str)
+}
+