+++ src/parser_rcpp_wrap.h
@@ -3,6 +3,7 @@
 
 #include "parse_option_ast.h"
 #include "parse_chunk_ast.h"
+#include "parse_code_block_ast.h"
 #include "parse_rmd_ast.h"
 
 #include <RcppCommon.h>
@@ -12,6 +13,8 @@ namespace Rcpp {
   template <> SEXP wrap(client::ast::chunk const& chunk);
   template <> SEXP wrap(std::vector<client::ast::chunk> const& chunks);
   template <> SEXP wrap(std::vector<client::ast::option> const& opts);
+
+  template <> SEXP wrap(client::ast::code_block const& block);
 
   // rmd
   template <> SEXP wrap(client::ast::heading const& h);