+++ tests/testthat/test-parse_rmd.R
@@ -57,11 +57,37 @@ test_that("knitr examples", {
     if (grepl("065-rmd-chunk\\.Rmd", file))
       next
 
-    label = paste("Parsing", basename(file))
-    expect_error(parse_rmd(!!file, allow_incomplete = FALSE), regexp = NA, label = label)
+    test_that(basename(file), {
+      ast = expect_no_error(parse_rmd(file), class = "rmd_ast")
+
+      expect_equal(
+        ast,
+        parse_rmd(as_document(ast, padding = character()))
+      )
+    })
   }
 })
 
+
+test_that("quarto web examples", {
+  files = dir(
+    system.file("tests/testthat/quarto-web/", package = "parsermd"),
+    pattern = utils::glob2rx("*.qmd"),
+    recursive = TRUE,
+    full.names = TRUE
+  )
+
+  for(file in files) {
+    test_that(basename(file), {
+      #ast = expect_no_error(parse_rmd(file), class = "rmd_ast")
+
+      #expect_equal(
+      #  ast,
+      #  parse_rmd(as_document(ast, padding = character()))
+      #)
+    })
+  }
+})
 
 
 test_that("Found issues", {